+++ src/_/css.ts
@@ -63,45 +63,6 @@ declare const _CssediaQueryString__Brand: unique symbol;
 export type CssMediaQueryString = string & { [_CssediaQueryString__Brand]: true };
 
 /**
- * @deprecated  Use DIY `@at-root` wraps instead  (Will be removed in v0.6+)
- *
- * This helper that emits a top-level `@media` block — i.e. one that forecully
- * @at-root breaks out any all ancestor `@media` blocks.
- *
- * This mixin is really only useful as a defence mechanism inside CSS-emitting
- * functions (mixins) that will be used in unpredictable `@media` contexts.
- */
-export function media(
-  query: CssMediaQueryString | string
-): (cssContent: CssString | string) => CssString;
-export function media(
-  query: CssMediaQueryString | string,
-  cssContent: CssString | string
-): CssString;
-
-/*#__NO_SIDE_EFFECTS__*/
-export function media(
-  query: CssMediaQueryString | string,
-  cssContent?: CssString | string
-) {
-  const mediaWrappinator = (cssContent: CssString | string) =>
-    css`
-      ${'@at-root (without: media)'} {
-        @media ${query} {
-          ${cssContent}
-        }
-      }
-    `;
-
-  if (arguments.length === 1) {
-    // Curry!
-    return mediaWrappinator;
-  }
-
-  return mediaWrappinator(cssContent || '');
-}
-
-/**
  * Helper to convert a value to a "quoted string".
  *
  * @see https://github.com/maranomynet/es-in-css/tree/v0.7#str-quoted-string-printer
@@ -113,9 +74,3 @@ export const str = (string: string) => JSON.stringify(`${string}`);
 // Deprecated types:
 
 declare const _RawCssString__Brand: unique symbol;
-/** @deprecated  Use `CssString` type instead (Will be removed in v0.8) */
-export type RawCssString = string & { [_RawCssString__Brand]?: true };
-/** @deprecated  Use either `RawCssString` or just `string` instead (Will be removed in v0.8) */
-export type RawCssValue = string & { [_RawCssString__Brand]?: true };
-/** @deprecated  Use `CssMediaQueryString` type instead (Will be removed in v0.8) */
-export type RawMediaQuery = string & { [_RawCssString__Brand]?: true };